Refactor author avatar URL generation
