+++ src/IHFiction.FictionApi/Authors/GetAuthor.cs
@@ -1,6 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.Security.Cryptography;
-using System.Text;
 using System.Runtime.Serialization;
 
 using Microsoft.AspNetCore.Mvc;
@@ -75,7 +73,7 @@ internal sealed class GetAuthor(FictionDbContext context) : IUseCase, INameEndpo
         return new GetAuthorResponse(
             author.UserId,
             author.Name,
-            BuildAvatarUrl(author.GravatarEmail),
+            GravatarUrlBuilder.BuildAvatarUrl(author.GravatarEmail),
             author.UpdatedAt,
             author.DeletedAt,
             new GaAuthorProfile(
@@ -85,28 +83,6 @@ internal sealed class GetAuthor(FictionDbContext context) : IUseCase, INameEndpo
                     .Select(link => new GaAuthorSocialLink(link.Type, link.Value))),
             publishedStories,
             totalStories);
-    }
-
-    private static string? BuildAvatarUrl(string? gravatarEmail)
-    {
-        if (string.IsNullOrWhiteSpace(gravatarEmail))
-        {
-            return null;
-        }
-
-        #pragma warning disable CA1308 // Gravatar canonicalization requires lowercase email and hash
-        var normalized = gravatarEmail.Trim().ToLowerInvariant();
-        #pragma warning restore CA1308
-
-        #pragma warning disable CA5351 // MD5 is required by Gravatar hashing protocol
-        var hashBytes = MD5.HashData(Encoding.UTF8.GetBytes(normalized));
-        #pragma warning restore CA5351
-
-        #pragma warning disable CA1308 // Gravatar canonicalization requires lowercase email and hash
-        var hash = Convert.ToHexString(hashBytes).ToLowerInvariant();
-        #pragma warning restore CA1308
-
-        return $"https://www.gravatar.com/avatar/{hash}?s=512&d=identicon&r=g";
     }
 
     public static string EndpointName => nameof(GetAuthor);
