+++ src/IHFiction.FictionApi/Authors/ListAuthors.cs
@@ -1,6 +1,4 @@
 using System.ComponentModel.DataAnnotations;
-using System.Security.Cryptography;
-using System.Text;
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -99,7 +97,7 @@ internal sealed class ListAuthors(
             .Select(a => new ListAuthorsItem(
                 a.Id,
                 a.Name,
-                BuildAvatarUrl(a.GravatarEmail),
+                GravatarUrlBuilder.BuildAvatarUrl(a.GravatarEmail),
                 a.Profile.Bio ?? "",
                 a.CreatedAt,
                 a.UpdatedAt,
@@ -109,28 +107,6 @@ internal sealed class ListAuthors(
         // Execute paginated query using the centralized service
         return await paginator.ExecutePagedQueryAsync(proj, query, cancellationToken);
     }
-
-        private static string? BuildAvatarUrl(string? gravatarEmail)
-        {
-        if (string.IsNullOrWhiteSpace(gravatarEmail))
-        {
-            return null;
-        }
-
-    #pragma warning disable CA1308 // Gravatar canonicalization requires lowercase email and hash
-        var normalized = gravatarEmail.Trim().ToLowerInvariant();
-    #pragma warning restore CA1308
-
-    #pragma warning disable CA5351 // MD5 is required by Gravatar hashing protocol
-        var hashBytes = MD5.HashData(Encoding.UTF8.GetBytes(normalized));
-    #pragma warning restore CA5351
-
-    #pragma warning disable CA1308 // Gravatar canonicalization requires lowercase email and hash
-        var hash = Convert.ToHexString(hashBytes).ToLowerInvariant();
-    #pragma warning restore CA1308
-
-        return $"https://www.gravatar.com/avatar/{hash}?s=512&d=identicon&r=g";
-        }
 
     public static string EndpointName => nameof(ListAuthors);
